requirements.txt
#client
python-socketio[client]
requests
time

#server
python-socketio
flask
cryptography
eventlet